refactor: 金額入力を共通化
月予算作成フォームの準備として、支払い機能で使っていたAmountInputを共通入力コンポーネントへ移動する。

Storyとテストも共通コンポーネント側へ移し、支払い機能側は既存import互換のre-exportにする。
