+++ apps/web/src/components/inputs/AmountInput/index.ts
@@ -1,0 +1,1 @@
+export * from "./AmountInput"
